Add op collection and op to library
