+++ tritonbench/__init__.py
@@ -1,3 +1,2 @@
-from . import operator_loader
 from . import operators
 from . import operators_collection
